fix(fcm): MB-2961, fcm token not unregistered fix

+++ src/framework/modules/auth/service.ts
@@ -35,6 +35,7 @@ import {
   UserChild,
   UserChildren,
 } from './model';
+import { getSession } from './reducer';
 
 import { I18n } from '~/app/i18n';
 import appConf, { Platform } from '~/framework/util/appConf';
@@ -486,21 +487,30 @@ export class FcmService {
 
   async unregisterFCMToken(token: string | null = null) {
     try {
+      const account = getSession();
+      if (!account) {
+        console.error('FcmService - unregisterFCMToken - ERROR - No account');
+        return;
+      }
       if (!token) {
         token = await messaging().getToken();
       }
       if (token) {
-        await signedFetch(`${this.platform.url}/timeline/pushNotif/fcmToken?fcmToken=${token}`, {
-          method: 'delete',
-        });
+        const req = OAuth2RessourceOwnerPasswordClient.signRequestWithToken(
+          OAuth2RessourceOwnerPasswordClient.convertTokenToOldObjectSyntax(account.tokens),
+          `${this.platform.url}/timeline/pushNotif/fcmToken?fcmToken=${token}`,
+          {
+            method: 'delete',
+          },
+        );
+        await fetch(req);
         this._removeTokenFromDeleteQueue(token);
         console.debug('FcmService - unregisterFCMToken - OK - ', token);
       } else {
         console.debug('FcmService - unregisterFCMToken - NO TOKEN - ');
       }
     } catch (err) {
-      console.error('FcmService - unregisterFCMToken - ERROR - ', token);
-      console.error((err as Error).message);
+      console.error('FcmService - unregisterFCMToken - ERROR - ', (err as Error).message);
       //unregistering fcm token should not crash the login process
       if (!Connection.isOnline) {
         //when no connection => get it from property
@@ -522,7 +532,6 @@ export class FcmService {
           method: 'delete',
         },
       );
-
       await fetch(req);
       this._removeTokenFromDeleteQueue(token);
       console.debug('FcmService - unregisterFCMTokenWithAccount - OK - ', token);
